Update project properties and enhance logging
Updated output type in project settings for BedrockSlim.Tests, Server.Sample and Client.Sample. Enhanced logging in MyCustomProcessor class to utilize ILogger instead of Console.WriteLine. Also, updated the method in the Worker class of Client.Sample to log the received response. Enhanced ServiceCollectionExtensions on the client side to allow configuration of TcpClient.